refactor: Se refactorizan todos los hooks y services
