+++ src/hooks/products/useCharacteristics.ts
