+++ src/hooks/specifications/useSpecificationsManagement.ts
