+++ src/hooks/specifications/useSpecs.ts
@@ -98,93 +98,3 @@ function useSpecs({ catId, initialSpecs }: Spec) {
 }
 
 export default useSpecs;
-
-/* import { SpecDataFrontend } from '@interfaces/Specifications.interface';
-import { fetchSpecificationsByCategoryId } from '@services/specifications/fetchSpecificationsByCategoryId.service';
-import { useEffect, useState } from 'react';
-
-interface Spec {
-	catId: string;
-	initialSpecs?: Array<{ specification: string; value: string }>;
-}
-
-function useSpecs({ catId, initialSpecs }: Spec) {
-	console.log(initialSpecs);
-	const [specs, setSpecs] = useState<Array<{ specification: string; value: string }>>(
-		initialSpecs || [],
-	);
-	const [specifications, setSpecifications] = useState<SpecDataFrontend[]>([]);
-	const [flag, setFlag] = useState(true);
-	const [specValues, setSpecValues] = useState<Record<string, string>>({});
-
-	const editOrCreateKeyInSpecs = (keyToFind: string, newValue: string) => {
-		let keyFound = false;
-		const updatedSpecs = specs.map(spec => {
-			if (spec.specification === keyToFind) {
-				keyFound = true;
-				return { ...spec, value: newValue };
-			}
-			return spec;
-		});
-
-		if (!keyFound) {
-			updatedSpecs.push({ specification: keyToFind, value: newValue });
-		}
-
-		setSpecs(updatedSpecs);
-	};
-
-	const findKeyInSpecs = (keyToFind: string) => {
-		const spec = specs.find(spec => spec.specification === keyToFind);
-		return spec ? spec.value : null;
-	};
-
-	const handleSpecUpdate = (newValue: string, index: number) => {
-		const key = specifications[index - 1]?.id;
-		if (key) {
-			editOrCreateKeyInSpecs(key, newValue);
-		}
-	};
-
-	useEffect(() => {
-		if (!flag) {
-			setSpecs([]);
-		} else {
-			setFlag(false);
-		}
-
-		const getSpecifications = async () => {
-			try {
-				const data = await fetchSpecificationsByCategoryId(catId);
-				setSpecifications(data);
-			} catch (error) {
-				console.error('No se pudieron obtener las especificaciones');
-			}
-		};
-
-		if (catId) {
-			getSpecifications();
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [catId]);
-
-	const handleInputChange = (id: string, value: string) => {
-		setSpecValues(prevValues => ({
-			...prevValues,
-			[id]: value,
-		}));
-	};
-
-	return {
-		specs,
-		specifications,
-		editOrCreateKeyInSpecs,
-		findKeyInSpecs,
-		handleSpecUpdate,
-		specValues,
-		handleInputChange,
-	};
-}
-
-export default useSpecs;
-*/
