+++ src/pages/specificationsPages/EditSpecification.tsx
@@ -8,8 +8,8 @@ import SuccessMessage from '@components/SuccessMessage';
 import TrashIcon from '@components/svg/TrashIcon';
 import useErrorHandling from '@hooks/common/useErrorHandling';
 import useSuccessHandling from '@hooks/common/useSuccessHandling';
-import useSpecificationsManagement from '@hooks/useSpecificationsManagement';
-import useToolCategorizationEdit from '@hooks/useToolCategorizationEdit';
+import useToolCategorizationEdit from '@hooks/products/useToolCategorizationEdit';
+import useSpecificationsManagement from '@hooks/specifications/useSpecificationsManagement';
 import { SpecDataToSend } from '@interfaces/Specifications.interface';
 import BasePage from '@layouts/BasePage';
 import HeaderApp from '@layouts/HeaderApp';
