+++ src/pages/toolPages/CreateTool.tsx
@@ -7,10 +7,10 @@ import { AuthContext } from '@contexts/AuthContext';
 import useErrorHandling from '@hooks/common/useErrorHandling';
 import useSuccessHandling from '@hooks/common/useSuccessHandling';
 import useMultipleFileUpload from '@hooks/files/useMultipleFileUpload';
-import useCharacteristics from '@hooks/useCharacteristics';
-import useSpecs from '@hooks/useSpecs';
-import useToolCategorizationEdit from '@hooks/useToolCategorizationEdit';
-import useVideos from '@hooks/useVideos';
+import useVideos from '@hooks/files/useVideos';
+import useCharacteristics from '@hooks/products/useCharacteristics';
+import useToolCategorizationEdit from '@hooks/products/useToolCategorizationEdit';
+import useSpecs from '@hooks/specifications/useSpecs';
 import BasePage from '@layouts/BasePage';
 import HeaderApp from '@layouts/HeaderApp';
 import '@pages/toolPages/createtool.css';
